src/requirements.txt
altgraph==0.17.2
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
future==0.18.2
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.2.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
keyboard==0.13.5
loguru==0.5.3
MarkupSafe==2.0.1
oauthlib==3.1.1
pefile==2021.9.3
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyparsing==2.4.7
python-dotenv==0.19.0
pywin32-ctypes==0.2.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.7
Werkzeug==2.0.1
win32-setctime==1.0.3
